% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hvk-package.R
\docType{package}
\name{hvk-package}
\alias{hvk}
\alias{hvk-package}
\title{hvk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
