% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read a file}
\usage{
read_file(
  path = NULL,
  type = c("txt", "csv", "tsv", "rds", "fst", "xlsx"),
  ...
)
}
\arguments{
\item{path}{a string}
}
\value{
a table
}
\description{
Read a file
}
